reciept_ocr/backend/requirements.txt
Flask
opencv-python
pytesseract
flask-cors
google-cloud-vision